+++ webapps-commons/src/main/java/com/gt/toolbox/spb/webapps/commons/infra/datamodel/LazyDMFiller.java
@@ -9,5 +9,5 @@ public interface LazyDMFiller<E> {
 
 	Page<E> findByFilter(Map<String, String> filters, Pageable pageable);
 
-	int countByFilter(Map<String, String> filters);
+	long countByFilter(Map<String, String> filters);
 }
